Move firebase into app/
