+++ src/slices/auth/auth-slice.ts
@@ -8,8 +8,8 @@ import {
   UserCredential,
 } from 'firebase/auth';
 import { getEnv } from '../../app/environment.js';
+import { auth as authApi, firebaseRefs } from '../../app/firebase.js';
 import { RootState, ThunkPromise, ThunkResult } from '../../app/store.js';
-import { auth as authApi, firebaseRefs } from '../../firebase.js';
 import { User } from '../../models/auth.js';
 import { logger } from '../../util/logger.js';
