+++ src/storage/firestore-writer.ts
@@ -13,8 +13,8 @@ import {
   updateDoc,
   WithFieldValue,
 } from 'firebase/firestore';
+import { firebaseRefs } from '../app/firebase.js';
 import { RootState } from '../app/store.js';
-import { firebaseRefs } from '../firebase.js';
 import { Model } from '../models/model.js';
 import { selectCurrentTeam } from '../slices/app/index.js';
 import { selectCurrentUserId } from '../slices/auth/index.js';
